Working on writing json to 2dp

+++ app/uk/gov/hmrc/lisaapi/models/ResorceModel.scala
@@ -96,19 +96,11 @@ case class LisaAccount(investorID:ID,
 
 case class TransferAccount(transferredFromAccountID:AccountId, transferredFromLMRN:ReferenceNumber, transferInDate:ISO8601Date)
 
-case class Bonuses(bonusDueForPeriod:Float,
-                   totalBonusDueYTD:Float,
-                   bonusPaidYTD:Option[Float],
-                   claimReason:String)
+case class Bonuses(bonusDueForPeriod: Double, totalBonusDueYTD: Double, bonusPaidYTD: Option[Double], claimReason: String)
 
-case class HelpToBuyTransfer(htbTransferInForPeriod:Float, htbTransferTotalYTD:Float)
+case class HelpToBuyTransfer(htbTransferInForPeriod: Double, htbTransferTotalYTD: Double)
 
-case class InboundPayments(
-  newSubsForPeriod: Option[Float],
-  newSubsYTD: Float,
-  totalSubsForPeriod: Float,
-  totalSubsYTD: Float
-)
+case class InboundPayments(newSubsForPeriod: Option[Double], newSubsYTD: Double, totalSubsForPeriod: Double, totalSubsYTD: Double)
 
 case class LifeEvent(accountID: AccountId,
                       lisaManagerReferenceNumber:ReferenceNumber,
